+++ CHANGELOG.md
@@ -34,3 +34,13 @@ All notable changes to this project will be documented in this file.
 - Parameter `n_neighbors` in `atlas.pp.preprocessing` updated to include None and adopt standard behavior for `mu.pp.neighbors`.
 - Parameter `knn_rna` and `knn_act` in `atlas.pp.neighbors` set to 15 as in `sc.pp.neighbors` default version 1.12.0.
 - Parameter `n_pcs_rna` and `n_pcs_act` in `atlas.pp.preprocessing` set to 50 as in `sc.pp.pca` default version 1.12.0.
+
+## [0.0.9] - 2026-04-20
+### Changed
+- Changed plot size in `atlas.pl.plot_trends` and adjusted for labels and legend.
+
+
+## [0.0.11] - 2026-05-18
+### Changed
+- Added `backend` parameter to atlas.tl.CellRankExtension.compute_transition_matrix. The parameter is used for parallelization.
+- Added `backend` parameter to atlas.tl.CellRankExtension.run. The parameter is used for parallelization.